lottie/expressions: force the precision down to single-precision floating point
This also removes some minor compiler warings with "Wdouble-promotion"

issue: https://github.com/thorvg/thorvg/issues/3136

+++ src/loaders/lottie/jerryscript/jerry-core/ecma/base/ecma-helpers-errol.cpp
@@ -170,7 +170,7 @@ ecma_errol0_dtoa (double val, /**< ecma number */
   high_bound.value = mid.value;
   high_bound.offset = mid.offset;
 
-  if (ECMA_NEXT_FLOAT (val) != INFINITY)
+  if ((float) ECMA_NEXT_FLOAT (val) != (float) INFINITY)
   {
     high_bound.offset += (ECMA_NEXT_FLOAT (val) - val) * power_of_10 / (2.0 + ERROL0_EPSILON);
   }
